# apps/api/openapi.json
{"openapi":"3.1.0","info":{"title":"Tianho Website API","description":"HTTP contract for the public website and administration area.","version":"v1"},"servers":[{"url":"http://127.0.0.1:8080","description":"Generated server url"}],"paths":{"/api/v1/auth/mfa/verify":{"post":{"tags":["administrator-session-controller"],"operationId":"verify","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VerificationResponse"}}}}}}},"/api/v1/auth/mfa/recovery":{"post":{"tags":["administrator-session-controller"],"operationId":"recover","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecoveryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VerificationResponse"}}}}}}},"/api/v1/auth/mfa/bind":{"post":{"tags":["administrator-session-controller"],"operationId":"bind","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChallengeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BindingResponse"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["administrator-session-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChallengeResponse"}}}}}}},"/api/v1/admin/{collection}/{id}/{action}":{"post":{"tags":["content-revision-controller"],"operationId":"changeState","parameters":[{"name":"collection","in":"path","required":true,"schema":{"type":"string"}},{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"action","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VersionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContentStateResponse"}}}}}}},"/api/v1/admin/users":{"get":{"tags":["administrator-user-controller"],"operationId":"users","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"post":{"tags":["administrator-user-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/v1/admin/media":{"get":{"tags":["media-library-controller"],"operationId":"findAssets","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","default":"ACTIVE","enum":["ACTIVE","TRASHED","DELETED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MediaAssetResponse"}}}}}}},"post":{"tags":["media-library-controller"],"operationId":"upload","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MediaAssetResponse"}}}}}}},"/api/v1/admin/media/{id}/trash":{"post":{"tags":["media-library-controller"],"operationId":"trash","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MediaAssetResponse"}}}}}}},"/api/v1/admin/media/{id}/restore":{"post":{"tags":["media-library-controller"],"operationId":"restore","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MediaAssetResponse"}}}}}}},"/api/v1/admin/media/{id}/purge":{"post":{"tags":["media-library-controller"],"operationId":"purge","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"No Content"}}}},"/api/v1/admin/articles":{"get":{"tags":["admin-content-controller"],"operationId":"articles","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/JsonNode"}}}}}}},"post":{"tags":["admin-content-controller"],"operationId":"createArticle","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleInput"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JsonNode"}}}}}}},"/api/v1/admin/works/{id}/images":{"patch":{"tags":["admin-content-controller"],"operationId":"updateWorkImages","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkImagesInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JsonNode"}}}}}}},"/api/v1/admin/users/{id}/status":{"patch":{"tags":["administrator-user-controller"],"operationId":"changeStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeStatusRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/v1/admin/users/{id}/role":{"patch":{"tags":["administrator-user-controller"],"operationId":"changeRole","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeRoleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/v1/admin/notices":{"get":{"tags":["admin-content-controller"],"operationId":"notices","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/JsonNode"}}}}}}},"patch":{"tags":["admin-content-controller"],"operationId":"saveNotice","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoticeInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JsonNode"}}}}}}},"/api/v1/admin/articles/{id}":{"get":{"tags":["admin-content-controller"],"operationId":"article","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JsonNode"}}}}}},"patch":{"tags":["admin-content-controller"],"operationId":"updateArticle","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JsonNode"}}}}}}},"/api/v1/public/works":{"get":{"tags":["public-content-controller"],"operationId":"works","parameters":[{"name":"locale","in":"query","required":false,"schema":{"type":"string","default":"ja","pattern":"ja|zh|en"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkSummary"}}}}}}}},"/api/v1/public/works/{slug}":{"get":{"tags":["public-content-controller"],"operationId":"work","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}},{"name":"locale","in":"query","required":false,"schema":{"type":"string","default":"ja","pattern":"ja|zh|en"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WorkDetail"}}}}}}},"/api/v1/public/notices/current":{"get":{"tags":["public-content-controller"],"operationId":"currentNotice","parameters":[{"name":"locale","in":"query","required":false,"schema":{"type":"string","default":"ja","pattern":"ja|zh|en"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Notice"}}}}}}},"/api/v1/public/media/{variant}/{filename}":{"get":{"tags":["public-media-controller"],"operationId":"read","parameters":[{"name":"variant","in":"path","required":true,"schema":{"type":"string"}},{"name":"filename","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}}}}},"/api/v1/public/articles":{"get":{"tags":["public-content-controller"],"operationId":"articles_1","parameters":[{"name":"locale","in":"query","required":false,"schema":{"type":"string","default":"ja","pattern":"ja|zh|en"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ArticleSummary"}}}}}}}},"/api/v1/public/articles/{slug}":{"get":{"tags":["public-content-controller"],"operationId":"article_1","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}},{"name":"locale","in":"query","required":false,"schema":{"type":"string","default":"ja","pattern":"ja|zh|en"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleDetail"}}}}}}},"/api/v1/auth/session":{"get":{"tags":["administrator-session-controller"],"operationId":"session","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SessionResponse"}}}}}}},"/api/v1/auth/csrf":{"get":{"tags":["administrator-session-controller"],"operationId":"csrf","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CsrfResponse"}}}}}}},"/api/v1/admin/{collection}/{id}/revisions":{"get":{"tags":["content-revision-controller"],"operationId":"findRevisions","parameters":[{"name":"collection","in":"path","required":true,"schema":{"type":"string"}},{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RevisionResponse"}}}}}}}},"/api/v1/admin/works":{"get":{"tags":["admin-content-controller"],"operationId":"works_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/JsonNode"}}}}}}}},"/api/v1/admin/works/{id}":{"get":{"tags":["admin-content-controller"],"operationId":"work_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JsonNode"}}}}}}},"/api/v1/admin/media/{id}":{"get":{"tags":["media-library-controller"],"operationId":"findAsset","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MediaAssetResponse"}}}}}}}},"components":{"schemas":{"VerifyRequest":{"type":"object","properties":{"challengeId":{"type":"string","format":"uuid"},"code":{"type":"string","minLength":1,"pattern":"\\d{6}"}},"required":["challengeId","code"]},"SessionResponse":{"type":"object","properties":{"authenticated":{"type":"boolean"},"userId":{"type":"string","format":"uuid"},"email":{"type":"string"},"displayName":{"type":"string"},"role":{"type":"string"}}},"VerificationResponse":{"type":"object","properties":{"session":{"$ref":"#/components/schemas/SessionResponse"},"recoveryCodes":{"type":"array","items":{"type":"string"}}}},"RecoveryRequest":{"type":"object","properties":{"challengeId":{"type":"string","format":"uuid"},"recoveryCode":{"type":"string","minLength":1,"pattern":"[A-Za-z2-7]{4}-[A-Za-z2-7]{4}-[A-Za-z2-7]{4}"}},"required":["challengeId","recoveryCode"]},"ChallengeRequest":{"type":"object","properties":{"challengeId":{"type":"string","format":"uuid"}},"required":["challengeId"]},"BindingResponse":{"type":"object","properties":{"secret":{"type":"string"},"provisioningUri":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"ChallengeResponse":{"type":"object","properties":{"challengeId":{"type":"string","format":"uuid"},"mfaRequired":{"type":"boolean"},"setupRequired":{"type":"boolean"}}},"VersionRequest":{"type":"object","properties":{"expectedVersion":{"type":"integer","format":"int64"}},"required":["expectedVersion"]},"ContentStateResponse":{"type":"object","properties":{"resourceType":{"type":"string","enum":["ARTICLE","WORK","NOTICE"]},"id":{"type":"string","format":"uuid"},"status":{"type":"string","enum":["DRAFT","PUBLISHED","ARCHIVED"]},"version":{"type":"integer","format":"int64"},"archivedAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"CreateUserRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"displayName":{"type":"string","maxLength":160,"minLength":0},"password":{"type":"string","maxLength":200,"minLength":16},"role":{"type":"string","enum":["ADMIN","EDITOR"]}},"required":["displayName","email","password","role"]},"UserResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"displayName":{"type":"string"},"role":{"type":"string","enum":["ADMIN","EDITOR"]},"active":{"type":"boolean"},"verifiedAt":{"type":"string","format":"date-time"},"lastLoginAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"}}},"MediaAssetResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"originalFilename":{"type":"string"},"contentType":{"type":"string"},"byteSize":{"type":"integer","format":"int64"},"width":{"type":"integer","format":"int32"},"height":{"type":"integer","format":"int32"},"sha256":{"type":"string"},"status":{"type":"string","enum":["ACTIVE","TRASHED","DELETED"]},"url":{"type":"string"},"thumbnailUrl":{"type":"string"},"referenceCount":{"type":"integer","format":"int64"},"trashedAt":{"type":"string","format":"date-time"},"purgeAfter":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"}}},"ArticleBlockInput":{"type":"object","properties":{"type":{"type":"string","minLength":1,"pattern":"heading|paragraph|image"},"heading":{"type":"string"},"body":{"type":"string"},"imagePath":{"type":"string"},"imageAlt":{"type":"string"},"imageTone":{"type":"string","minLength":1,"pattern":"neutral|warm|cool|rust"},"caption":{"type":"string"},"sortOrder":{"type":"integer","format":"int32","minimum":0}},"required":["imageTone","type"]},"ArticleInput":{"type":"object","properties":{"locale":{"type":"string","minLength":1,"pattern":"ja|zh|en"},"slug":{"type":"string","minLength":1,"pattern":"[a-z0-9]+(?:-[a-z0-9]+)*"},"title":{"type":"string","minLength":1},"excerpt":{"type":"string","minLength":1},"category":{"type":"string","minLength":1},"authorName":{"type":"string","minLength":1},"heroLabel":{"type":"string"},"heroImagePath":{"type":"string"},"heroAlt":{"type":"string"},"heroTone":{"type":"string","minLength":1,"pattern":"neutral|warm|cool|rust"},"heroCaption":{"type":"string"},"closingNote":{"type":"string"},"ctaLabel":{"type":"string"},"ctaHref":{"type":"string"},"displayOrder":{"type":"integer","format":"int32","minimum":0},"relatedServices":{"type":"array","items":{"type":"string","minLength":1}},"seoTitle":{"type":"string"},"seoDescription":{"type":"string"},"youtubeUrl":{"type":"string"},"demo":{"type":"boolean"},"blocks":{"type":"array","items":{"$ref":"#/components/schemas/ArticleBlockInput"},"maxItems":2147483647,"minItems":1}},"required":["authorName","blocks","category","excerpt","heroTone","locale","relatedServices","slug","title"]},"JsonNode":{"type":"object","properties":{"empty":{"type":"boolean"},"array":{"type":"boolean"},"null":{"type":"boolean"},"object":{"type":"boolean"},"float":{"type":"boolean"},"number":{"type":"boolean"},"int":{"type":"boolean"},"binary":{"type":"boolean"},"string":{"type":"boolean"},"pojo":{"type":"boolean"},"double":{"type":"boolean"},"boolean":{"type":"boolean"},"short":{"type":"boolean"},"long":{"type":"boolean"},"textual":{"type":"boolean","deprecated":true},"missingNode":{"type":"boolean"},"container":{"type":"boolean"},"bigInteger":{"type":"boolean"},"valueNode":{"type":"boolean"},"floatingPointNumber":{"type":"boolean"},"integralNumber":{"type":"boolean"},"bigDecimal":{"type":"boolean"},"nodeType":{"type":"string","enum":["ARRAY","BINARY","BOOLEAN","MISSING","NULL","NUMBER","OBJECT","POJO","STRING"]},"embeddedValue":{"type":"boolean"}}},"WorkImageInput":{"type":"object","properties":{"path":{"type":"string","minLength":1},"label":{"type":"string","minLength":1},"tone":{"type":"string","minLength":1,"pattern":"neutral|warm|cool|rust"},"altJa":{"type":"string"},"altZh":{"type":"string"},"altEn":{"type":"string"},"captionJa":{"type":"string"},"captionZh":{"type":"string"},"captionEn":{"type":"string"},"isCover":{"type":"boolean"},"sortOrder":{"type":"integer","format":"int32","minimum":0}},"required":["label","path","tone"]},"WorkImagesInput":{"type":"object","properties":{"expectedVersion":{"type":"integer","format":"int64"},"galleryEnabled":{"type":"boolean"},"mediaType":{"type":"string","minLength":1,"pattern":"photo|gallery|video"},"images":{"type":"array","items":{"$ref":"#/components/schemas/WorkImageInput"}}},"required":["images","mediaType"]},"ChangeStatusRequest":{"type":"object","properties":{"active":{"type":"boolean"}}},"ChangeRoleRequest":{"type":"object","properties":{"role":{"type":"string","enum":["ADMIN","EDITOR"]}},"required":["role"]},"NoticeInput":{"type":"object","properties":{"locale":{"type":"string","minLength":1,"pattern":"ja|zh|en"},"enabled":{"type":"boolean"},"label":{"type":"string","minLength":1},"title":{"type":"string","minLength":1},"body":{"type":"string","minLength":1},"dismissLabel":{"type":"string","minLength":1},"linkLabel":{"type":"string"},"linkHref":{"type":"string"},"storageKey":{"type":"string","minLength":1},"dismissalMode":{"type":"string","minLength":1,"pattern":"session|local"},"startAt":{"type":"string","format":"date-time"},"endAt":{"type":"string","format":"date-time"},"expectedVersion":{"type":"integer","format":"int64"}},"required":["body","dismissLabel","dismissalMode","label","locale","storageKey","title"]},"ArticleUpdateRequest":{"type":"object","properties":{"expectedVersion":{"type":"integer","format":"int64"},"article":{"$ref":"#/components/schemas/ArticleInput"}},"required":["article"]},"WorkSummary":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"locale":{"type":"string"},"slug":{"type":"string"},"title":{"type":"string"},"summary":{"type":"string"},"category":{"type":"string"},"serviceCategory":{"type":"string"},"featuredOnHomepage":{"type":"boolean"},"featuredOrder":{"type":"integer","format":"int32"},"mediaType":{"type":"string"},"coverImagePath":{"type":"string"},"coverImageAlt":{"type":"string"},"coverImageTone":{"type":"string"}}},"WorkDetail":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"translationGroupId":{"type":"string","format":"uuid"},"locale":{"type":"string"},"slug":{"type":"string"},"title":{"type":"string"},"summary":{"type":"string"},"clientName":{"type":"string"},"projectDate":{"type":"string"},"category":{"type":"string"},"serviceCategory":{"type":"string"},"scope":{"type":"string"},"challenge":{"type":"string"},"approach":{"type":"array","items":{"type":"string"}},"outcome":{"type":"string"},"deliverables":{"type":"array","items":{"type":"string"}},"featuredOnHomepage":{"type":"boolean"},"featuredOrder":{"type":"integer","format":"int32"},"mediaType":{"type":"string"},"galleryEnabled":{"type":"boolean"},"seoTitle":{"type":"string"},"seoDescription":{"type":"string"},"youtubeUrl":{"type":"string"},"version":{"type":"integer","format":"int64"},"updatedAt":{"type":"string","format":"date-time"},"images":{"type":"array","items":{"$ref":"#/components/schemas/WorkImage"}}}},"WorkImage":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"path":{"type":"string"},"label":{"type":"string"},"tone":{"type":"string"},"alt":{"type":"string"},"caption":{"type":"string"},"cover":{"type":"boolean"},"sortOrder":{"type":"integer","format":"int32"}}},"Notice":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"translationGroupId":{"type":"string","format":"uuid"},"locale":{"type":"string"},"label":{"type":"string"},"title":{"type":"string"},"body":{"type":"string"},"dismissLabel":{"type":"string"},"linkLabel":{"type":"string"},"linkHref":{"type":"string"},"storageKey":{"type":"string"},"dismissalMode":{"type":"string"},"startAt":{"type":"string","format":"date-time"},"endAt":{"type":"string","format":"date-time"},"version":{"type":"integer","format":"int64"}}},"ArticleSummary":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"locale":{"type":"string"},"slug":{"type":"string"},"title":{"type":"string"},"excerpt":{"type":"string"},"category":{"type":"string"},"authorName":{"type":"string"},"heroImagePath":{"type":"string"},"heroAlt":{"type":"string"},"heroTone":{"type":"string"},"publishedAt":{"type":"string","format":"date-time"},"displayOrder":{"type":"integer","format":"int32"},"demo":{"type":"boolean"}}},"ArticleBlock":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"type":{"type":"string"},"heading":{"type":"string"},"body":{"type":"string"},"imagePath":{"type":"string"},"imageAlt":{"type":"string"},"imageTone":{"type":"string"},"caption":{"type":"string"},"sortOrder":{"type":"integer","format":"int32"}}},"ArticleDetail":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"translationGroupId":{"type":"string","format":"uuid"},"locale":{"type":"string"},"slug":{"type":"string"},"title":{"type":"string"},"excerpt":{"type":"string"},"category":{"type":"string"},"authorName":{"type":"string"},"heroLabel":{"type":"string"},"heroImagePath":{"type":"string"},"heroAlt":{"type":"string"},"heroTone":{"type":"string"},"heroCaption":{"type":"string"},"closingNote":{"type":"string"},"ctaLabel":{"type":"string"},"ctaHref":{"type":"string"},"publishedAt":{"type":"string","format":"date-time"},"relatedServices":{"type":"array","items":{"type":"string"}},"seoTitle":{"type":"string"},"seoDescription":{"type":"string"},"youtubeUrl":{"type":"string"},"demo":{"type":"boolean"},"version":{"type":"integer","format":"int64"},"updatedAt":{"type":"string","format":"date-time"},"blocks":{"type":"array","items":{"$ref":"#/components/schemas/ArticleBlock"}}}},"CsrfResponse":{"type":"object","properties":{"headerName":{"type":"string"},"parameterName":{"type":"string"},"token":{"type":"string"}}},"RevisionResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"version":{"type":"integer","format":"int64"},"action":{"type":"string"},"snapshot":{"$ref":"#/components/schemas/JsonNode"},"actorId":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"}}}}}}